+++ NEWS.md
@@ -1,3 +1,11 @@
+# gtR 0.3.0
+
+* Function that encompasses and replaces all previous query functions (`get_resources()`).
+
+# gtR 0.2.0
+
+* Functions to run queries against a combination of resources (`query_resource_combination()`), and a list of resource combinations (`gtr_combinations`).
+
 # gtR 0.1.0
 
 * Functions to get API configuration for each resource type (`get_configs()`), and generic function to run queries against each resource(`query_resource_all()`.